+++ app/src/main/java/com/example/mealio/view/mainScreen/star/StarFragment.java
@@ -1,23 +1,29 @@
 package com.example.mealio.view.mainScreen.star;
 
+import android.content.Intent;
 import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
-import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
+
 import com.example.mealio.R;
 import com.example.mealio.model.MealRepository;
 import com.example.mealio.model.Network.MealRemoteDataSourceImp;
 import com.example.mealio.model.db.Meal;
 import com.example.mealio.model.db.MealLocalDataSourceImp;
 import com.example.mealio.presenter.StarPresenter;
+import com.example.mealio.view.credentialScreen.AuthenticationActivity;
 import com.example.mealio.view.mainScreen.Home.HomeFragmentDirections;
 import com.example.mealio.view.mainScreen.Home.OnMealClickListener;
+import com.example.mealio.view.mainScreen.Utils;
 import com.google.android.material.snackbar.Snackbar;
 import java.util.List;
 
@@ -52,8 +58,8 @@ public class StarFragment extends Fragment implements OnDeleteClickListener, Sta
         starPresenter = new StarPresenter(this, MealRepository.getInstance(MealLocalDataSourceImp.getInstance(getContext()),
                 MealRemoteDataSourceImp.getInstance()));
 
-        GridLayoutManager gridLayoutManager = new GridLayoutManager(getContext(), 2);
-        starMealsRecyclerView.setLayoutManager(gridLayoutManager);
+       LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getActivity());
+        starMealsRecyclerView.setLayoutManager(linearLayoutManager);
         starMealsRecyclerView.setAdapter(starMealAdapter);
 
         starPresenter.getStoredMeal();
@@ -62,10 +68,18 @@ public class StarFragment extends Fragment implements OnDeleteClickListener, Sta
 
     @Override
     public void deleteFromStar(Meal meal) {
-        starPresenter.deleteFromFav(meal);
-        starMealAdapter.notifyDataSetChanged();
-        Snackbar.make(starMealsRecyclerView,"Meal Was deleted successfully" , Snackbar.LENGTH_LONG).show();
+        new AlertDialog.Builder(requireContext())
+                .setTitle("Confirm Deletion")
+                .setMessage("Are you sure you want to delete this meal?")
+                .setPositiveButton("Yes", (dialog, which) -> {
+                    starPresenter.deleteFromFav(meal);
+                    starMealAdapter.notifyDataSetChanged();
+                    Snackbar.make(starMealsRecyclerView, "Meal was deleted successfully", Snackbar.LENGTH_LONG).show();
+                })
+                .setNegativeButton("No", (dialog, which) -> dialog.dismiss())
+                .show();
     }
+
 
     @Override
     public void setMeals(List<Meal> meals) {
@@ -79,8 +93,20 @@ public class StarFragment extends Fragment implements OnDeleteClickListener, Sta
 
     @Override
     public void OnMealClicked(String id) {
-        StarFragmentDirections.ActionStarFragmentToMealDetailsFragment action =
-                StarFragmentDirections.actionStarFragmentToMealDetailsFragment(id,"StarFragment", "DetailsFragment");
-        Navigation.findNavController(starMealsRecyclerView).navigate(action);
+
+        if (Utils.isNetworkAvailable(requireContext())) {
+            StarFragmentDirections.ActionStarFragmentToMealDetailsFragment action =
+                    StarFragmentDirections.actionStarFragmentToMealDetailsFragment(id,"StarFragment","MealDetailsFragment");
+
+            Navigation.findNavController(starMealsRecyclerView).navigate(action);
+        } else {
+            StarFragmentDirections.ActionStarFragmentToStarMealDetailsFragment action =
+                    StarFragmentDirections.actionStarFragmentToStarMealDetailsFragment(id);
+
+            Navigation.findNavController(starMealsRecyclerView).navigate(action);
+        }
+
     }
+
+
 }
